Common test file for shared functions
